Initialize spaceship position only in game
