+++ crates/lumina_server/src/lobby/in_game.rs
@@ -6,13 +6,13 @@ use lumina_shared::prelude::*;
 use lumina_terrain::prelude::*;
 use server::*;
 
-use super::{LobbyFull, LobbyInGame, LobbySeed};
+use super::{LobbyFull, LobbyInGame, LobbyInfos, LobbySeed};
 
 pub(super) struct InGamePlugin;
 
 impl Plugin for InGamePlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(Update, start_game);
+        app.add_systems(Update, (start_game, init_spaceship_position));
     }
 }
 
@@ -42,3 +42,43 @@ fn start_game(
         commands.entity(entity).insert(LobbyInGame);
     }
 }
+
+/// This function updates the position of newly spawned spaceships
+fn init_spaceship_position(
+    mut commands: Commands, // Need Commands to add Position if missing
+    mut q_spaceships: Query<
+        (&mut Position, &PlayerId, Entity),
+        (
+            With<Spaceship>,
+            With<SourceEntity>,
+            Without<PositionInitialized>,
+        ),
+    >,
+    q_in_game_lobbies: Query<(), With<LobbyInGame>>,
+    terrain_config: TerrainConfig,
+    lobby_infos: Res<LobbyInfos>,
+) {
+    let Some(terrain_config) = terrain_config.get() else {
+        return;
+    };
+
+    // TODO: Use different positions for different ships based on their team id.
+    let width = terrain_config.tile_size * terrain_config.noise_surr_width as f32;
+    let desired_position = Vec2::splat(width);
+
+    for (mut position, id, entity) in q_spaceships.iter_mut() {
+        if lobby_infos
+            .get(&**id)
+            .is_some_and(|e| q_in_game_lobbies.contains(*e))
+            == false
+        {
+            continue;
+        }
+
+        *position = Position(desired_position);
+        commands.entity(entity).insert(PositionInitialized);
+    }
+}
+
+#[derive(Component)]
+pub struct PositionInitialized;
